oneoff/requirements.txt
boto3==1.11.15
botocore==1.14.15
Brotli==1.0.9
docutils==0.15.2
entrypoints==0.3
flake8==3.7.9
jmespath==0.9.4
jsonlines==1.2.0
mccabe==0.6.1
mysql-connector-python==8.0.21
protobuf==3.13.0
psycopg2==2.8.4
pycodestyle==2.5.0
pyflakes==2.1.1
PyPika==0.35.21
python-dateutil==2.8.1
s3transfer==0.3.3
six==1.14.0
urllib3==1.25.8
